CAP331: Fix to the reports page search filters.
Before, the reports page search filters weren't working as expected,
limitations on the presence on a non null value of dates made it non
flexible. Also, fields like the reportNo and the report type were added,
so that needed to make them part of the search as well.

This change comes to make search filters more flexible but precise.
Pagination was disabled for now but will be added back soon.